- added hash header with fnv1 - added profiler - actually working frfr hash table - added spirv_reflect - added StaticStr<N> - fixed arena bug where cleanup was called on empty arena

+++ src/gfx/descriptor.h
@@ -51,7 +51,7 @@ struct DescriptorLayoutCache {
 
         bool operator==(const DescriptorLayoutInfo &other) const;
 
-        usize hash() const;
+        u32 hash() const;
     };
 
 private:
